[all]: cleanup

+++ TankLibTest/Program.cs
@@ -248,12 +248,12 @@ namespace TankLibTest {
         
         public static void TestTexturePayload() {
             teResourceGUID guid = (teResourceGUID)0xC00000000000A00;
-            teResourceGUID payloadGuid = (teResourceGUID)(((ulong)guid & 0xF0FFFFFFFFUL) | 0x100000000UL | 0x0320000000000000UL);
+            teResourceGUID payloadGuid = (teResourceGUID)((guid & 0xF0FFFFFFFFUL) | 0x100000000UL | 0x0320000000000000UL);
             teTexture texture;
-            using (Stream textureStream = OpenFile((ulong) guid)) {
+            using (Stream textureStream = OpenFile(guid)) {
                 texture = new teTexture(textureStream);
             }
-            using (Stream texturePayloadStream = OpenFile((ulong) payloadGuid)) {
+            using (Stream texturePayloadStream = OpenFile(payloadGuid)) {
                 texture.LoadPayload(texturePayloadStream);
             }
             
@@ -264,7 +264,7 @@ namespace TankLibTest {
 
         public static void TestTexture() {
             teResourceGUID guid = (teResourceGUID)0xC0000000000C708;
-            using (Stream textureStream = OpenFile((ulong) guid)) {
+            using (Stream textureStream = OpenFile(guid)) {
                 teTexture texture = new teTexture(textureStream);
                 using (Stream file = File.OpenWrite(guid+".dds")) {
                     texture.SaveToDDS(file);
@@ -342,7 +342,7 @@ namespace TankLibTest {
             using (Stream matStream = OpenFile(guid)) {
                 material = new teMaterial(matStream);
 
-                using (Stream matDataStream = OpenFile((ulong) material.Header.MaterialData)) {
+                using (Stream matDataStream = OpenFile(material.Header.MaterialData)) {
                     materialData = new teMaterialData(matDataStream);
                 }
             }
